test: Split overly large util_tests.cpp file
